WIP: seprate session control
